Removes array to object casting; more flexible, consistent with docs.
PUT, POST and PATCH methods currently have documentation and exceptions
indicating that both objects and arrays can be provided, but in each
case any provided objects will be cast into an array.

The default behaviour of the AbtstractRestfulController, and therefore
the PhlyRestfully/ResourceController, is to unencode json data as it's
provided, so arrays are decoded as arrays and objects as objects.

Sticking with this, rather than always modifying to an object, allows
more flexibility in the listeners (e.g. listening for an array for
create to create a collection of resources, vs an object to create just
one).