Supports bulk update and/or create of resources through patchList.
Implements support for the AbstractRrestfulController::patchList method,
which should be used for updating and/or creating a collection of
resources.

Implementation is very similar to replaceList, as the replace vs patch
logic is largely implemented in the listeners.

Tests are implemented based on the replaceList tests, with some small
extensions to both to more fully test the validity of arguments.