Allow the default controller class to be specified.
Currently each resource can have a separate controller class specified,
but there is no way to specify a default controller class to use where
this is not specified; it's hard coded to
PhlyRestfully\ResourceController.

This patch allows a global default controller class to be specified and
defaults to PhlyRestfully\ResourceController if this is not specified.